미션 시스템 구현 - 텐트치기, 요리하기

+++ Assets/06.Scripts/Exp/UITextExp.cs
@@ -10,7 +10,7 @@ public class UITextExp : MonoBehaviour
     // Start is called before the first frame update
     void Start()
     {
-        ExpManager.Instance.Subscribe_Exp((float exp) =>
+        ExpManager.Instance.Subscribe_Exp((int exp) =>
         {
             text.text = $"{exp}";
         });
